large scale refactor
