+++ arbiter-core/src/database/inspector.rs
