+++ arbiter-core/src/middleware/transaction.rs
@@ -2,8 +2,7 @@ use revm::primitives::{ExecutionResult, Output};
 
 /// Unwraps the result of the EVM execution into a more structured `Success`
 /// type.
-use super::cast::revm_logs_to_ethers_logs;
-use super::errors::RevmMiddlewareError;
+use super::*;
 
 /// Contains the result of a successful transaction execution.
 #[derive(Debug)]
